+++ anytime/src/main/java/com/naver/anytime/mybatis/mapper/PostLikeMapper.java
@@ -1,14 +1,13 @@
 package com.naver.anytime.mybatis.mapper;
 
 import org.apache.ibatis.annotations.Mapper;
-import org.apache.ibatis.annotations.Update;
-
 import com.naver.anytime.domain.PostLike;
 
 @Mapper
 public interface PostLikeMapper {
     
 	public void addNewlike(PostLike postLike);
+	
     public void removeExistinglike(PostLike postLike);
     public int checkIfUserAlreadyLiked(PostLike postLike);
