+++ anytime/src/main/java/com/naver/anytime/mybatis/mapper/TimeTable_detailMapper.java
@@ -4,7 +4,6 @@ import java.util.List;
 
 import org.apache.ibatis.annotations.Mapper;
 
-import com.naver.anytime.domain.TimeTable;
 import com.naver.anytime.domain.TimeTable_detail;
 
 @Mapper
