+++ anytime/src/main/java/com/naver/anytime/service/TimeTable_detailService.java
@@ -2,7 +2,6 @@ package com.naver.anytime.service;
 
 import java.util.List;
 
-import com.naver.anytime.domain.TimeTable;
 import com.naver.anytime.domain.TimeTable_detail;
 
 public interface TimeTable_detailService {
