+++ anytime/src/main/resources/static/js/post/postPhoto.js
@@ -23,7 +23,7 @@ $(document).ready(function () {
 			        $('.vote').text(data.like_count); // 서버로부터 받은 공감 수를 표시
 			    } else if(data.statusCode == 2){
 			        alert('공감 취소!');
-			        $('.vote').text(data.like_count); // 서버로부터 받은 공감 수를 표시
+			        $('.vote').text(data.like_count); 
 			    } else{
 			        alert('오류 발생 : ' + data.errorMessage);
 			    }
