+++ anytime/src/main/resources/static/js/post/write.js
@@ -1,7 +1,7 @@
 $(document).ready(function(){
 
 	$("#writeBoardContainer").prepend(
-	`<form class="write" id ="writeBoard">
+	`<form class="write" id ="writeBoard" enctype="multipart/form-data">
 	
 	<p><input id="title" placeholder="글 제목" class="title"></p>
 	<p style="margin-bottom: 0px !important;">
@@ -32,7 +32,9 @@ $(document).ready(function(){
 </textarea></p>
 <input class="file" type="file" name="file" multiple="multiple">
 <ol class="thumbnails">
-<li class="new"></li>
+<li class="new">
+	<input type="file" id="new_file" name="file" multiple style='display:none;'>
+</li>
 </ol>
 <div class="clearBothOnly"></div>
 <ul class="option">
@@ -69,51 +71,113 @@ $("#writeArticleButton").on("click", function(e) {
 	if(toggle == true){
 		
 		$("#writeBoard").show(); //Show
+		$('#writeArticleButton').hide();
 	}else{		
 		
 		$("#writeBoard").hide(); // Hide
+		$('#writeArticleButton').show();
 	}
 });
 
 /*글쓰기 완료*/
-  $("#writePost").on("click", function(e) {
-	e.preventDefault(); //form태그 action안타게 막아버리는것
-	
-	let token = $("meta[name='_csrf']").attr("content");
-    let header = $("meta[name='_csrf_header']").attr("content");
-	
-	var board_id = document.getElementById('board_id').textContent;
-	
-	console.log($("#title").val());
-	console.log($("#content").val()); 
-	
-  	let data = {
-		BOARD_ID: board_id,
-		LOGIN_ID: $('#login_id').val(),
-		SUBJECT: $("#title").val(),
-		CONTENT: $("#content").val(),
-	};
-      
-   	console.log(data);
-   	
-      $.ajax({
-		type: "POST",
-		url:"write",
-		data: data,
-		async: false,
-		beforeSend: function(xhr) {
-				xhr.setRequestHeader(header, token);
-      		},
-	}).done(function(res) {
-    if(res.statusCode == 1){
-        alert("게시글 작성에 성공하였습니다.");
-        location.reload();
-    } else {
-        alert("게시글 작성에 실패하였습니다. 오류: " + res.errorMessage);
-    }
-});
-      
-});
+var filesToUpload = []; // 업로드할 파일들을 저장하는 배열
 
+$('body').on('click', 'li.new', function() {
+	    $('#post_file').trigger('click');
+	});
+
+	$('#post_file, #new_file').on('change', function() {
+		var newFiles = this.files;
+	
+		// 이미 업로드된 파일의 개수
+		var uploadedFileCount = $('.thumbnails li').length - 1; //'new' li 제외
+	
+		if(uploadedFileCount + newFiles.length > 5) {
+			alert("최대 5개의 파일만 첨부 가능합니다.");
+			this.value=''; // input 태그의 value 초기화
+			return;
+		}
+	
+		for(var i=0; i<newFiles.length; i++){
+			if(!newFiles[i].type.startsWith('image/')){
+				alert('현재 지원되지 않는 형식이거나, \n파일이 손상되었기 때문에 사진을 열 수 없습니다.');
+				this.value='';
+				continue;
+			}
+	
+			filesToUpload.push(newFiles[i]); // 새롭게 추가된 파일들도 배열에 포함시킴
+	
+			var reader = new FileReader();
+	
+			reader.onloadend = function(e) {
+				// 여기서 썸네일 미리보기를 생성합니다.
+				var thumbnailItem=$('<li><img src="' + e.target.result + '" style="width:85px; height:85px;"></li>');
+				
+				$('.thumbnails .new').before(thumbnailItem);
+				
+				// 첨부파일이 선택되었으므로, thumbnails 영역을 보여줍니다.
+				$('#container>div.articles>form.write ol.thumbnails').css('display', 'block');
+			};
+			
+			reader.readAsDataURL(newFiles[i]);
+		}
+		
+		// input 태그의 value 초기화
+		this.value='';
+	});
+
+	$('body').on('click', '.thumbnails li:not(.new)', function() {
+		var index = $(this).index();
+		
+		// 썸네일 삭제
+		$(this).remove();
+	
+		// 배열에서 파일제거
+		filesToUpload.splice(index, 1);
+	
+		 // 업로드된 파일의 개수가 0이면 thumbnails 영역을 숨깁니다.
+		 if($('.thumbnails li').length == 1) { //'new' li만 남았다면
+			 $('#container>div.articles>form.write ol.thumbnails').css('display', 'none');
+		 }
+	});
+
+
+    $("#writePost").on("click", function(e) {
+        e.preventDefault();
+
+        let token = $("meta[name='_csrf']").attr("content");
+        let header = $("meta[name='_csrf_header']").attr("content");
+
+	    var board_id =$("#board_id").html(); 
+
+	    var formData=new FormData();
+	    formData.append('BOARD_ID', board_id);
+	    formData.append('LOGIN_ID', $('#login_id').val());
+	    formData.append('SUBJECT', $("#title").val());
+	    formData.append('CONTENT', $("#content").val());
+
+	    for (var i=0; i<filesToUpload.length; i++) { 
+		    // 모든 첨부파일을 formData에 추가합니다.
+		    formData.append("file[]", filesToUpload[i]);
+	    }
+
+	  	$.ajax({
+	    	type: "POST",
+	    	url: "/anytime/post/write",
+	    	data: formData,
+	    	processData: false,  
+	  		contentType: false,
+	  		beforeSend: function(xhr) {
+	  			xhr.setRequestHeader(header, token);
+	  		},
+	  	}).done(function(res) {
+	      	if (res.statusCode == 1){
+	          	alert("게시글 작성에 성공하였습니다.");
+	          	location.reload();
+	      	} else {
+	          	alert("게시글 작성에 실패하였습니다. 오류: " + res.errorMessage);
+	      	}
+	  	});
+    });
 /*글쓰기 완료*/
 })
