+++ anytime/src/main/webapp/WEB-INF/views/main/anytimeMain.jsp
@@ -5,6 +5,7 @@
 <!DOCTYPE html>
 <html>
 <head>
+<link rel="shortcut icon" type="image/x-icon" href="${pageContext.request.contextPath}/resources/image/common/nav.logo.png">
 <title>애니타임</title>
 <link href="${pageContext.request.contextPath}/resources/css/main/main.css" type="text/css" rel="stylesheet">
