+++ anytime/src/main/webapp/WEB-INF/views/member/agreementForm.jsp
@@ -2,6 +2,7 @@
 <!DOCTYPE html>
 <html>
 <head>
+<link rel="shortcut icon" type="image/x-icon" href="${pageContext.request.contextPath}/resources/image/common/nav.logo.png">
 <title>애니타임</title>
 <script src="http://code.jquery.com/jquery-1.11.0.min.js"></script>
 <link type="text/css"
