+++ anytime/src/main/webapp/WEB-INF/views/member/loginForm.jsp
@@ -4,6 +4,7 @@
 <!DOCTYPE html>
 <html>
 <head>
+<link rel="shortcut icon" type="image/x-icon" href="${pageContext.request.contextPath}/resources/image/common/nav.logo.png">
 <title>애니타임</title>
 <link
 	href="${pageContext.request.contextPath}/resources/css/common/common.css"
