+++ anytime/src/main/webapp/WEB-INF/views/member/registerForm.jsp
@@ -3,6 +3,7 @@
 <!DOCTYPE html>
 <html>
 <head>
+<link rel="shortcut icon" type="image/x-icon" href="${pageContext.request.contextPath}/resources/image/common/nav.logo.png">
 <title>애니타임</title>
 <link type="text/css"
 	rel="stylesheet" href="${pageContext.request.contextPath}/resources/css/member/join/register.css" >
